Move some more common code out.
